functions/requirements.txt
azure-functions==1.18.0
azure-servicebus==7.11.4
requests==2.31.0
pydantic==2.5.2